Update athom-smart-plug.yaml
It seems that the actual chip used en PG03 is not HLW8012 but BL0937. I haven't open the plug, but it is apparent from the readings.
In that case, using the new `model` option, the multipliers are correctly computed and no linear filter is needed.
This way all measures are correct, while the current implementation shows an incorrect `energy` if it is used directly (instead of the integration).
I calculated the voltage multiplier empirically. You could provide the actual resistors value.